# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Metawall API",
    "description": "Metawall API 文件"
  },
  "host": "localhost:3005",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "headers",
      "name": "Authorization",
      "description": "JWT Token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/meta/user/register": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "註冊",
        "description": "",
        "parameters": [
          {
            "name": "request.body",
            "in": "body",
            "required": true,
            "description": "註冊資料",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "a1234567"
                },
                "nickName": {
                  "type": "string",
                  "example": "暱稱"
                }
              },
              "required": [
                "email",
                "password",
                "nickName"
              ]
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "註冊成功",
            "schema": {
              "$ref": "#/definitions/Sign"
            }
          },
          "400": {
            "description": "註冊失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/meta/user/login": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "登入",
        "description": "",
        "parameters": [
          {
            "name": "request.body",
            "in": "body",
            "required": true,
            "description": "登入資料",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@gmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "a1234567"
                }
              }
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "登入成功",
            "schema": {
              "$ref": "#/definitions/Sign"
            }
          },
          "400": {
            "description": "登入失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/meta/user/forgetPassword": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "忘記密碼",
        "description": "",
        "parameters": [
          {
            "name": "request.body",
            "in": "body",
            "required": true,
            "description": "忘記密碼資料",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@gmail.com"
                }
              },
              "required": [
                "email"
              ]
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/Mailer"
            }
          },
          "400": {
            "description": "失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/meta/user/verification/{userId}": {
      "post": {
        "tags": [
          "Login"
        ],
        "summary": "驗證碼",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "request.body",
            "in": "body",
            "required": true,
            "description": "驗證碼資料",
            "schema": {
              "type": "object",
              "properties": {
                "verification": {
                  "type": "string",
                  "example": "12345"
                }
              },
              "required": [
                "verification"
              ]
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "驗證成功",
            "schema": {
              "$ref": "#/definitions/Sign"
            }
          },
          "400": {
            "description": "驗證失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/meta/user/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "修改密碼",
        "description": "",
        "parameters": [
          {
            "name": "request.body",
            "in": "body",
            "required": true,
            "description": "修改密碼資料",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "a1234567"
                },
                "confirmPassword": {
                  "type": "string",
                  "example": "a1234567"
                }
              },
              "required": [
                "password",
                "confirmPassword"
              ]
            },
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "修改密碼成功",
            "schema": {
              "$ref": "#/definitions/Sign"
            }
          },
          "400": {
            "description": "修改密碼失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/meta/user/profile": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "取得個人頁面",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "取得個人資料成功",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "取得訂單列表失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "編輯個人資料",
        "description": "",
        "parameters": [
          {
            "name": "request.body",
            "in": "body",
            "required": true,
            "description": "編輯個人資料",
            "schema": {
              "type": "object",
              "properties": {
                "nickName": {
                  "type": "string",
                  "example": "test"
                },
                "gender": {
                  "type": "string",
                  "example": "male"
                }
              },
              "required": [
                "nickName",
                "gender"
              ]
            },
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "編輯個人資料成功",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "編輯個人資料失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/meta/user/profile/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "取得他人個人頁面",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "取得個人資料成功",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "取得訂單列表失敗",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "無此路由",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "系統錯誤",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "錯誤訊息"
        }
      }
    },
    "Sign": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string",
          "example": "abcde"
        },
        "nickname": {
          "type": "string",
          "example": "test"
        }
      }
    },
    "Mailer": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "請至 Email 查收信件"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "data": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "62bd81280e4fd7c1a022ad29"
            },
            "nickName": {
              "type": "string",
              "example": "sss"
            },
            "gender": {
              "type": "string",
              "example": "male"
            },
            "avatar": {
              "type": "string",
              "example": ""
            },
            "createdAt": {
              "type": "string",
              "example": "2022-06-30T10:55:36.985Z"
            },
            "updateAt": {
              "type": "string",
              "example": "2022-06-30T10:55:36.986Z"
            },
            "followers": {
              "type": "array",
              "items": {}
            },
            "followings": {
              "type": "array",
              "items": {}
            }
          }
        }
      }
    },
...
  },
...
}